moved all pre-checks and related code to a separate file
